backend/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
asttokens==3.0.1
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
boto3==1.42.8
botocore==1.42.8
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.3
dagshub-annotation-converter==0.1.15
databricks-sdk==0.73.0
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
dnspython==1.16.0
docker==7.1.0
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.124.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.61.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huey==2.5.5
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
mlflow-tracing==3.7.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
nvidia-nccl-cu12==2.28.9
opentelemetry-api==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
packaging==24.2
pandas==2.3.3
parso==0.8.5
pathvalidate==3.3.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pymongo==3.12.0
pyparsing==3.2.5
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sendgrid==6.12.5
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.3.0
tornado==6.5.2
traitlets==5.14.3
treelib==1.8.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uv==0.9.17
uvicorn==0.38.0
wcwidth==0.2.14
Werkzeug==3.1.4
xgboost==3.1.2
yarl==1.22.0
zipp==3.23.0
